Added more unit tests, xmldoc and general code cleanup. Version bumped to 8.14.0
