+++ OnixLabs.Core/Result.Success.cs
@@ -82,8 +82,8 @@ public sealed class Success : Result, IValueEquatable<Success>
     /// Executes the function that matches the value of the current <see cref="Result"/> instance and returns its result.
     /// <remarks>In the case of a success, the success branch is invoked.</remarks>
     /// </summary>
-    /// <param name="success">The action to execute when the current <see cref="Result"/> instance is in a successful state.</param>
-    /// <param name="failure">The action to execute when the current <see cref="Result"/> instance is in a failed state.</param>
+    /// <param name="success">The function to execute when the current <see cref="Result"/> instance is in a successful state.</param>
+    /// <param name="failure">The function to execute when the current <see cref="Result"/> instance is in a failed state.</param>
     /// <typeparam name="TResult">The underlying type of the result produced by the matching function.</typeparam>
     /// <returns>
     /// Returns the result of the <paramref name="success"/> function if the current <see cref="Result"/> instance is in a successful state;
@@ -108,7 +108,7 @@ public sealed class Success : Result, IValueEquatable<Success>
     /// <param name="selector">The function to apply to the current <see cref="Result"/> instance.</param>
     /// <typeparam name="TResult">The underlying type of the result produced by the selector function.</typeparam>
     /// <returns>
-    /// Returns <see cref="Success{T}"/> if the current <see cref="Result"/> is in a successful state, and the action invocation is also successful; otherwise; <see cref="Failure{T}"/>.
+    /// Returns <see cref="Success{T}"/> if the current <see cref="Result"/> is in a successful state, and the function invocation is also successful; otherwise; <see cref="Failure{T}"/>.
     /// </returns>
     public override Result<TResult> Select<TResult>(Func<TResult> selector) => Result<TResult>.Of(selector);
 
@@ -118,7 +118,7 @@ public sealed class Success : Result, IValueEquatable<Success>
     /// </summary>
     /// <param name="selector">The function to apply to the current <see cref="Result"/> instance.</param>
     /// <returns>
-    /// Returns <see cref="Success"/> if the current <see cref="Result"/> is in a successful state, and the action invocation is also successful; otherwise; <see cref="Failure"/>.
+    /// Returns <see cref="Success"/> if the current <see cref="Result"/> is in a successful state, and the function invocation is also successful; otherwise; <see cref="Failure"/>.
     /// </returns>
     public override Result SelectMany(Func<Result> selector) => selector();
 
@@ -129,7 +129,7 @@ public sealed class Success : Result, IValueEquatable<Success>
     /// <param name="selector">The function to apply to the current <see cref="Result"/> instance.</param>
     /// <typeparam name="TResult">The underlying type of the result produced by the selector function.</typeparam>
     /// <returns>
-    /// Returns <see cref="Success{T}"/> if the current <see cref="Result"/> is in a successful state, and the action invocation is also successful; otherwise; <see cref="Failure{T}"/>.
+    /// Returns <see cref="Success{T}"/> if the current <see cref="Result"/> is in a successful state, and the function invocation is also successful; otherwise; <see cref="Failure{T}"/>.
     /// </returns>
     public override Result<TResult> SelectMany<TResult>(Func<Result<TResult>> selector) => selector();
 
@@ -244,8 +244,8 @@ public sealed class Success<T> : Result<T>, IValueEquatable<Success<T>>
     /// Executes the function that matches the value of the current <see cref="Result{T}"/> instance and returns its result.
     /// <remarks>In the case of a success, the success branch is invoked.</remarks>
     /// </summary>
-    /// <param name="success">The action to execute when the current <see cref="Result{T}"/> instance is in a successful state.</param>
-    /// <param name="failure">The action to execute when the current <see cref="Result{T}"/> instance is in a failed state.</param>
+    /// <param name="success">The function to execute when the current <see cref="Result{T}"/> instance is in a successful state.</param>
+    /// <param name="failure">The function to execute when the current <see cref="Result{T}"/> instance is in a failed state.</param>
     /// <typeparam name="TResult">The underlying type of the result produced by the matching function.</typeparam>
     /// <returns>
     /// Returns the result of the <paramref name="success"/> function if the current <see cref="Result{T}"/> instance is in a successful state;
@@ -270,7 +270,7 @@ public sealed class Success<T> : Result<T>, IValueEquatable<Success<T>>
     /// <param name="selector">The function to apply to the current <see cref="Result"/> instance.</param>
     /// <typeparam name="TResult">The underlying type of the result produced by the selector function.</typeparam>
     /// <returns>
-    /// Returns <see cref="Success{T}"/> if the current <see cref="Result"/> is in a successful state, and the action invocation is also successful; otherwise; <see cref="Failure{T}"/>.
+    /// Returns <see cref="Success{T}"/> if the current <see cref="Result"/> is in a successful state, and the function invocation is also successful; otherwise; <see cref="Failure{T}"/>.
     /// </returns>
     public override Result<TResult> Select<TResult>(Func<T, TResult> selector) => Result<TResult>.Of(() => selector(Value));
 
@@ -280,7 +280,7 @@ public sealed class Success<T> : Result<T>, IValueEquatable<Success<T>>
     /// </summary>
     /// <param name="selector">The function to apply to the current <see cref="Result"/> instance.</param>
     /// <returns>
-    /// Returns <see cref="Success"/> if the current <see cref="Result"/> is in a successful state, and the action invocation is also successful; otherwise; <see cref="Failure"/>.
+    /// Returns <see cref="Success"/> if the current <see cref="Result"/> is in a successful state, and the function invocation is also successful; otherwise; <see cref="Failure"/>.
     /// </returns>
     public override Result SelectMany(Func<T, Result> selector) => selector(Value);
 
@@ -291,7 +291,7 @@ public sealed class Success<T> : Result<T>, IValueEquatable<Success<T>>
     /// <param name="selector">The function to apply to the current <see cref="Result"/> instance.</param>
     /// <typeparam name="TResult">The underlying type of the result produced by the selector function.</typeparam>
     /// <returns>
-    /// Returns <see cref="Success{T}"/> if the current <see cref="Result"/> is in a successful state, and the action invocation is also successful; otherwise; <see cref="Failure{T}"/>.
+    /// Returns <see cref="Success{T}"/> if the current <see cref="Result"/> is in a successful state, and the function invocation is also successful; otherwise; <see cref="Failure{T}"/>.
     /// </returns>
     public override Result<TResult> SelectMany<TResult>(Func<T, Result<TResult>> selector) => selector(Value);
