+++ OnixLabs.Security.Cryptography.UnitTests/MerkleTreeTests.cs
@@ -36,6 +36,9 @@ public sealed class MerkleTreeTests
         // Then
         Assert.Equal(a, b);
         Assert.Equal(a.Hash, b.Hash);
+        Assert.True(a.Equals(b));
+        Assert.True(a == b);
+        Assert.False(a != b);
     }
 
     [Fact(DisplayName = "Different Merkle trees should not be considered equal")]
@@ -48,6 +51,9 @@ public sealed class MerkleTreeTests
         // Then
         Assert.NotEqual(a, b);
         Assert.NotEqual(a.Hash, b.Hash);
+        Assert.False(a.Equals(b));
+        Assert.False(a == b);
+        Assert.True(a != b);
     }
 
     [Fact(DisplayName = "MerkleTree.GetLeafHashes should produce the same leaf hashes that the tree was constructed with")]
